Add Splashscreen animation Restructure packages in root directory
