+++ app/src/main/java/com/arcanium/walletwatch/MainActivity.kt
@@ -1,15 +1,19 @@
 package com.arcanium.walletwatch
 
+import android.animation.ObjectAnimator
 import android.os.Bundle
+import android.view.View
+import android.view.animation.AnticipateInterpolator
 import androidx.activity.ComponentActivity
 import androidx.activity.compose.setContent
 import androidx.activity.viewModels
 import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.material3.MaterialTheme
 import androidx.compose.material3.Surface
-import androidx.compose.runtime.collectAsState
 import androidx.compose.ui.Modifier
+import androidx.core.animation.doOnEnd
 import androidx.core.splashscreen.SplashScreen.Companion.installSplashScreen
+import com.arcanium.walletwatch.core.presentation.MainViewModel
 import com.arcanium.walletwatch.ui.theme.WalletWatchTheme
 
 class MainActivity : ComponentActivity() {
@@ -18,13 +22,8 @@ class MainActivity : ComponentActivity() {
 
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
+        displaySplashScreen()
         viewModel.setUp()
-        installSplashScreen().apply {
-            setKeepOnScreenCondition {
-                viewModel.uiState.value.showSplashScreen
-            }
-        }
-
         setContent {
             WalletWatchTheme {
                 // A surface container using the 'background' color from the theme
@@ -34,6 +33,28 @@ class MainActivity : ComponentActivity() {
                 ) {
 
                 }
+            }
+        }
+    }
+
+    private fun displaySplashScreen() {
+        installSplashScreen().apply {
+            setKeepOnScreenCondition {
+                viewModel.uiState.value.showSplashScreen
+            }
+            setOnExitAnimationListener { splashScreenView ->
+                val slideUp = ObjectAnimator.ofFloat(
+                    splashScreenView.view,
+                    View.TRANSLATION_Y,
+                    0f,
+                    -splashScreenView.view.height.toFloat()
+                )
+                slideUp.interpolator = AnticipateInterpolator()
+                slideUp.duration = 400L
+
+                // Call SplashScreenView.remove at the end of animation.
+                slideUp.doOnEnd { splashScreenView.remove() }
+                slideUp.start()
             }
         }
     }
