+++ app/src/main/java/com/arcanium/walletwatch/core/presentation/MainActivityUiState.kt
@@ -1,4 +1,4 @@
-package com.arcanium.walletwatch
+package com.arcanium.walletwatch.core.presentation
 
 data class MainActivityUiState(
     val showSplashScreen: Boolean = true
